requirements.txt
seaborn
statsmodels
pyro-ppl
pandas
numpy
torch
matplotlib
scikit-learn